refactor: 파일 이름, 함수 리팩토링
